requirements.txt
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
branca==0.6.0
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-plugins==1.1.1
cligj==0.7.2
comm==0.1.4
command-not-found==0.3
contourpy==1.1.0
cramjam==2.7.0
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
distro==1.7.0
distro-info===1.1build1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fake-useragent==1.2.1
fastparquet==2023.7.0
Fiona==1.9.4.post1
Flask==2.3.3
folium==0.14.0
fonttools==4.42.0
fsspec==2023.6.0
geopandas==0.13.2
googlemaps==4.10.0
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
ipykernel==6.25.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.9.3
mapclassify==2.6.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
more-itertools==8.10.0
nest-asyncio==1.5.7
netifaces==0.11.0
networkx==3.1
numpy==1.25.2
oauthlib==3.2.0
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==1.5.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==12.0.1
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyproj==3.6.0
PySocks==1.7.1
pyspark==3.4.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==5.4.1
pyzmq==25.1.0
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
SecretStorage==3.3.1
selenium==4.12.0
shapely==2.0.1
six==1.16.0
sklearn==0.0.post7
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
ssh-import-id==5.11
stack-data==0.6.2
statsmodels==0.14.0
systemd-python==234
threadpoolctl==3.2.0
tornado==6.3.2
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.4
wadllib==1.3.6
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.7
wsproto==1.2.0
yellowbrick==1.5
zipp==1.0.0
